feat(outlook): fix messageId prop
