+++ packages/pieces/community/microsoft-outlook/src/lib/actions/add-label-to-email.ts
@@ -1,7 +1,7 @@
-import { createAction, Property, OAuth2PropertyValue } from '@activepieces/pieces-framework';
-import { Client, PageCollection } from '@microsoft/microsoft-graph-client';
-import { Message } from '@microsoft/microsoft-graph-types';
+import { createAction, Property } from '@activepieces/pieces-framework';
+import { Client } from '@microsoft/microsoft-graph-client';
 import { microsoftOutlookAuth } from '../common/auth';
+import { messageIdDropdown } from '../common/props';
 
 export const addLabelToEmailAction = createAction({
 	auth: microsoftOutlookAuth,
@@ -9,47 +9,10 @@ export const addLabelToEmailAction = createAction({
 	displayName: 'Add Label to Email',
 	description: 'Adds a category (label) to an email message.',
 	props: {
-		messageId: Property.Dropdown({
+		messageId: messageIdDropdown({
 			displayName: 'Email',
 			description: 'Select the email message to add the label to.',
 			required: true,
-			refreshers: [],
-			options: async ({ auth }) => {
-				if (!auth) {
-					return {
-						disabled: true,
-						options: [],
-					};
-				}
-
-				const client = Client.initWithMiddleware({
-					authProvider: {
-						getAccessToken: () => Promise.resolve((auth as OAuth2PropertyValue).access_token),
-					},
-				});
-
-				try {
-					const response: PageCollection = await client
-						.api('/me/messages?$top=50&$select=id,subject,from,receivedDateTime')
-						.orderby('receivedDateTime desc')
-						.get();
-
-					const messages = response.value as Message[];
-
-					return {
-						disabled: false,
-						options: messages.map((message) => ({
-							label: `${message.subject || 'No Subject'} - ${message.from?.emailAddress?.name || message.from?.emailAddress?.address || 'Unknown Sender'}`,
-							value: message.id || '',
-						})),
-					};
-				} catch (error) {
-					return {
-						disabled: true,
-						options: [],
-					};
-				}
-			},
 		}),
 		categories: Property.Array({
 			displayName: 'Categories',
